Move the SelectLanguagePage component into the Steps directory
